allow passing questions as props
